[refactor] 팀 동기화 시 영속성 컨텍스트 오염 문제 해결 - 서비스 분리로 각 팀 저장을 독립적인 트랜잭션으로 처리 - TeamService에 saveOrUpdate 메서드 이동 및 개별 트랜잭션 처리 - 에러 발생 시 영속성 컨텍스트 재사용으로 인한 문제 해결
